Don't require principals to be declared beforehand
Previously, Verifpal would error out if a model was written like this:

```
principal Alice[
    ...
]

Alice -> Bob: x

principal Bob[
    ...
]
```

This is because Verifpal would see that Alice was sending a message to
Bob before Bob is declared as a principal. This would require an
annoying pointless rewrite like this:

```
principal Alice[
    ...
]

principal Bob[]

Alice -> Bob: x

principal Bob[
    ...
]
```

With this change, we can keep the earlier model, and Verifpal will only
produce an error if Bob is never declared.