New: `leaks` keyword
As suggested by Tom Hacohen in
https://lists.symbolic.software/pipermail/verifpal/2020/000146.html:

"One more piece of syntax that I found a bit counter-intuitive, or at
least, not semantic is how PFS is checked. "Alice -> Bob: a", while
conveys the idea that Alice's private key has been leaked, it also
specifically indicates that Alice sends Bob the key as part of the
protocol, which is not the case. The more likely case is that Eve stole
Alice's device. Maybe a "becomes public a" or "leaks a" would be better
(those are just suggestions, better find something that you feel is more
consistent with the rest of verifpal)."

Tom is right, and so the `leaks` keyword now exists, allowing principals
to leak constants to the attacker without necessarily sending them to
another principal, like this:

principal Alice[
    knows private verySecretSecret
    leaks verySecretSecret
]

The attacker now knows verySecretSecret.

Support for the feature has been added to the VSCode/Vim extensions, and
it will also be documented in the User Manual by the end of the week.