Remove some mutex locks
At face value, this seems like a dangerous and bad thing to do: I am
removing mutex locks in two situations:

1. When checking to see if all queries have been contradicted/whether we
can terminate model verification, I do not lock the verifyResults struct
before reading from it.

2. Before reading from the attackerState struct to see whether the
attacker already knows a value, I do not lock the state immediately.
Rather, I only lock it if the value is not found in the attackerState,
*check again while locked to see if the value is still absent*, and then
write while locked.

The reason why I've made these changes (the latter change is far more
important than the former) is the following:

For change 1, there is simply no risk that can emerge from reading a
copy of verifyResults that is not the latest one. If not all queries are
contradicted, we'll just continue analyzing. So no harm done. On the
other hand, constantly locking verifyResults every single time we run an
analysis thread was costing precious parallelization opportunities for
no real benefit.

For change 2, there was truly an *enormous* impact on parallelization
that was happening from the fact that no analysis thread could obtain
even a casual idea of whether the attacker knew some value without
locking the attackerState away from all other analysis threads -- and
this is something that each analysis thread will imperatively need to
check frequently. And the thing is, if an initial check shows that the
attacker doesn't know a value, it's really fine: since a secondary check
does occur while attackerState is locked (and since, of course, no write
occurs on an unlocked attackerState), then we're fine. All we end up
doing here is limiting locked reads on attackerState with the purpose of
checking to see if a value has been used to *after* a check has occured
on a possibly slightly outdated attackerState. The worst that can happen
from a race condition here is that the analysis thread will do a
second, locked read check for nothing. Which is, again, in no way
harmful.

For these reasons, I am not convinced that these potential race conditions
present any risk whatsoever to Verifpal's analysis, and I am allowing
them to occur under controlled conditions in order to benefit from the
huge parallelization gains that I've repeatedly benchmarked and observed
in various standard model testing conditions.