Core primitives (with CONCAT and SPLIT)
Here we introduce the notion of "core primitives", which are simply
primitives in Verifpal that perform basic operations that are not
necessarily cryptographic.

From an engineering perspective as well as from a formal perspective,
this differentiation is useful because we can allow core primitives
to have more arbitrarily defined logic than the very rigidly specified
cryptographic primitives, which have to follow a primitiveSpec struct
in their specification.

The three current core primitives are:

1. ASSERT: used to be a cryptographic primitive, is now a core primitive.
Works same as before.

2. CONCAT: Provides concatenation, a very widely requested feature. c =
CONCAT(a, b). Supports up to five inputs.

3. SPLIT: Splits concatenated values. a, b = SPLIT(CONCAT(a, b)). Must have
CONCAT as its input, otherwise Verifpal will return an error and refuse
to run the analysis.

CONCAT and SPLIT support is currently experimental and might be buggy. I
have made some tests of course, but strongly encourage further testing.

The Verifpal User Manual has been updated to document core primitives.

The Verifpal syntax highlighting plugins for Visual Studio Code and Vim
have been updated to support the new CONCAT and SPLIT keywords.