Reorganize project structure into "lib" internals
In order to accomodate the fact that Verifpal will now be translating to
Coq, ProVerif and soon Go, the project structure inside the `internal`
folder is now such that we have Verifpal's code itself and a bunch of
"lib" folders:

- libpeg implements the Verifpal grammar in Peg and generates the
Verifpal model parser.
- libcoq implements the Verifpal Coq library/header.
- libpv implements the Verifpal ProVerif headers.
- libgo will implement the Verifpal Go implementation headers etc.

Whenever a "lib" is compiled, the resulting artifact goes in
internal/verifpal/libname.go.

The Makefile should hopefully compile these libs whenever is necessary
automatically.