Verifpal did not detect Lowe's Needham-Schroeder Attack
This fixes an issue where Verifpal failed to detect Gavin Lowe's attack
on the public-key variant of Needham-Schroeder.

In order for this attack to be detected, it should be only necessary to
model standard Needham-Schroeder and to unguard Bob's public key, while
leaving Alice's guarded. However, Verifpal would in this scenario not
detect the attack if Alice were to used a checked ASSERT in order to
verify that the message she received from Bob did indeed contain her
correct nonce on top of Bob's nonce. This is incorrect behavior.

The issue lay with the fact that the Verifpal mutation map did not
actually include a fully resolved expression of the original un-mutated
value by default, but only a pristine copy of the non-resolved
expression.

A new test has been added that shows that Verifpal can correctly detect
Lowe's attack in a standard Needham-Schroeder model with all the correct
checked ASSERTs and no undue "nudging" of Verifpal. A subsequent commit
will also contain a test that shows that Lowe's proposed fix for the
attack does indeed fix the attack in Verifpal.

Incidentally, the fix for this issue also dramatically speeds up
Verifpal's performance, cutting integration test run times by half.