Increase maximum stack size
The default Go stack size limit is insufficient for some highly complex
Verifpal analyses. I have tried to determine whether the stack size was
being reached due to an infinite loop and evidence suggests that this is
not the case. Therefore, the stack size limit is increased from the
default 1GB to 8GB.