Modify behavior of freshness queries
Now, freshness queries will check if a value is ever *used* while
non-fresh, and not simply contradict the query if a value can ever *be*
non-fresh. This is based on discussions with Eric Simpson.

An example model in examples/test/replay-simple.vp is provided to
illustrate how this is supposed to work.