Fix equivalent values resolution query overrides
This fixes an issue pointed out by "Melanie" on the Discord chat.
Prior to this fix, Verifpal would mark the query
`confidentiality? m2_b` in `examples/test/dh_equiv.vp` as passed,
because it resolves `m2_b` as equivalent to `m2` before checking to
see if the query is contradicted.

Thanks to Georgio Nicolas for assistance with debugging this issue.